refactor(server): share k8s template loader
